+++ client/src/Routes.js
@@ -4,4 +4,5 @@ export const LOGOUT_ROUTE = "/logout";
 export const COMPANIES_ROUTE = "/companies";
 export const MENTIONS_ROUTE = "/mentions";
 export const DIALOG_ROUTE = MENTIONS_ROUTE + "/mention/";
+export const JOBS_ROUTE = "/jobs/requests";
 export const SITES_ROUTE = "/settings/site/";
