+++ server/mentionscrawler/blueprints/__init__.py
@@ -3,6 +3,7 @@ from .session import session_bp
 from .site import site_bp
 from .mention import mention_bp
 from .company import company_bp
+from .job import job_bp
 
 
 # array of blueprints to be added to the app
