+++ server/mentionscrawler/crawlers/constants.py
@@ -1,0 +1,8 @@
+REDDIT = "Reddit"
+TWITTER = "Twitter"
+FACEBOOK = "Facebook"
+
+MENTIONS_TAG = "mentions"
+SECRET_HASH_TAG = "secret_hash"
+RESPONSE_URL = "/mentions/responses"
+SCHEDULE_TIME = 1  # time between crawls in minutes
